backend/requirements.txt
django
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
python-decouple
django-filter
pillow
django-cors-headers
django-redis
redis
boto3==1.28.0
botocore==1.31.0
django-storages
pytest
pytest-django
pytest-cov
coverage